Add files via upload
Handwritten Digit Recognition
In this project, we are going to implement a handwritten digit recognition app using the MNIST dataset. We will be using a special type of deep neural network that is Convolutional Neural Network (CNN). In the end, we are going to build a GUI in which you can draw the digit and recognize it straight away, and also accuracy of digit recognized is displayed in percentage. The digit can be handwritten using cursor/pointer on the GUI and the digit entered can be predicted.